requirements.txt
# PyTorch & Torchvision
torch>=2.0.0
torchvision>=0.15.0

# Data Processing and Scientific Computing
numpy>=1.21.0
pandas>=1.3.0

# Image processing
Pillow>=9.0.0

# Visualisation
matplotlib>=3.5.0

# A progress bar shows when training the model
tqdm>=4.60.0